Prepare for automatic semi-colon injection.
Semicolons are going to be part of the formal grammar of the
language. However, it is a chore to remember to feed the code with
this symbol and, occasionally, it might hurt the simplicity for
end user.

Following the same rules for BCPL and Go, the semicolons will be
injected automatically in the end of each line break, as long the
previous token is not part of an incomplete statement.