Invert relation between VM and Frontend.
Frontend, which is the last step in the chain, provided to VM a
pair of channels to communicate. It created an ordering problem in
which the VM, not being executed in background, did not allow for
frontends to be executed last. This is especially important for
go-qml, which expects its main loop to be tied with main.Main().